undoing stuff

+++ src/util.c
@@ -57,7 +57,7 @@ void print_properties_h(void) {
 	LOG(("  %f <=   mass   <= %f\n", massHeavyMin, massHeavyMax));
 }
 
-void print_particle(double *p, int time) {
+void print_particle(double *p, double time) {
 /*
 	LOG(("{\n"));
 	LOG(("\tparticle_id: %f,\n", p[ID_COL]));
@@ -70,7 +70,7 @@ void print_particle(double *p, int time) {
 	LOG(("}\n"));
 */
 /*
-	LOG(("t=%d, p_id=%d, mass=%f, p_x=%f, p_y=%f, v_x=%f, v_y=%f\n",
+	LOG(("t=%f, p_id=%d, mass=%f, p_x=%f, p_y=%f, v_x=%f, v_y=%f\n",
 			time, (int)p[ID_COL], p[WEIGHT_COL], p[POS_X_COL], p[POS_Y_COL],
 			p[VOL_X_COL], p[VOL_Y_COL]));
 */
@@ -78,14 +78,14 @@ void print_particle(double *p, int time) {
 	if (p[WEIGHT_COL] <= 0.0) {
 		return;
 	}
-	LOG(("%d,%d,%lf,%lf,%lf,%lf,%lf\n",
+	LOG(("%lf,%d,%lf,%lf,%lf,%lf,%lf\n",
 			time, (int)p[ID_COL], p[WEIGHT_COL], p[POS_X_COL], p[POS_Y_COL],
 			p[VOL_X_COL], p[VOL_Y_COL]));
 */
 	return;
 }
 
-void print_all_particles(double **P, int numParticle, int time) {
+void print_all_particles(double **P, int numParticle, double time) {
 	int i;
 	//LOG(("{\n"));
 	//LOG(("t,p_id,mass,p_x,p_y,v_x,v_y\n"));
